refactor: move interaction logic from generic to non-generic picker bases
All template wiring, event handlers, and state machine logic now lives in
the non-generic bases (DatePickerBase, DateTimePickerBase, DateRangePickerBase).
Abstract methods act as T-specific placeholders implemented by the generic bases.

- DatePickerBase: added OnApplyTemplate, all event handlers, 5 abstract methods
- DatePickerBase<T>: slimmed to SelectedDateProperty + converter abstracts + overrides
- DateTimePickerBase: added OnApplyTemplate, all event handlers, 7 abstract methods
- DateTimePickerBase<T>: slimmed to SelectedDateProperty + converter abstracts + overrides
- DateRangePickerBase: added OnApplyTemplate, full state machine, 6 abstract methods
- DateRangePickerBase<T>: slimmed to SelectedStart/EndDateProperty + converter abstracts + overrides
