+++ src/Ursa/Controls/DateTimePicker/DatePickerBaseT.cs
@@ -1,28 +1,13 @@
 using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Controls.Metadata;
 using Avalonia.Controls.Primitives;
 using Avalonia.Data;
-using Avalonia.Input;
-using Avalonia.Interactivity;
 using Irihi.Avalonia.Shared.Contracts;
-using Irihi.Avalonia.Shared.Helpers;
 
 namespace Ursa.Controls;
 
-[TemplatePart(PART_Popup, typeof(Popup))]
-[TemplatePart(PART_TextBox, typeof(TextBox))]
-[TemplatePart(PART_Calendar, typeof(DatePickerCalendarView))]
 public abstract class DatePickerBase<T> : DatePickerBase, IClearControl where T : struct
 {
-    public const string PART_Popup = "PART_Popup";
-    public const string PART_TextBox = "PART_TextBox";
-    public const string PART_Calendar = "PART_Calendar";
-
-    private TextBox? _textBox;
-    private DatePickerCalendarView? _calendar;
-    private Popup? _popup;
-
     public static readonly StyledProperty<T?> SelectedDateProperty =
         AvaloniaProperty.Register<DatePickerBase<T>, T?>(
             nameof(SelectedDate), defaultBindingMode: BindingMode.TwoWay);
@@ -48,127 +33,29 @@ public abstract class DatePickerBase<T> : DatePickerBase, IClearControl where T 
     /// <summary>Returns today as a <see cref="DateOnly"/> for use as the calendar context when no date is selected.</summary>
     protected virtual DateOnly GetToday() => DateOnly.FromDateTime(DateTime.Today);
 
-    protected override void OnPropertyChanged(AvaloniaPropertyChangedEventArgs change)
+    protected override DateOnly? GetSelectedDateOnly() =>
+        SelectedDate.HasValue ? ToDateOnly(SelectedDate.Value) : null;
+
+    protected override DateOnly GetCalendarContextDate() =>
+        GetSelectedDateOnly() ?? GetToday();
+
+    protected override void SyncToUI()
     {
-        base.OnPropertyChanged(change);
-        if (change.Property == SelectedDateProperty)
-            SyncSelectedDateToText(change.GetNewValue<T?>());
-    }
-
-    protected override void OnApplyTemplate(TemplateAppliedEventArgs e)
-    {
-        base.OnApplyTemplate(e);
-        DatePickerCalendarView.DateSelectedEvent.RemoveHandler(OnDateSelected, _calendar);
-        GotFocusEvent.RemoveHandler(OnTextBoxGotFocus, _textBox);
-        LostFocusEvent.RemoveHandler(OnTextBoxLostFocus, _textBox);
-        PointerPressedEvent.RemoveHandler(OnTextBoxPressed, _textBox);
-
-        _popup = e.NameScope.Find<Popup>(PART_Popup);
-        _textBox = e.NameScope.Find<TextBox>(PART_TextBox);
-        _calendar = e.NameScope.Find<DatePickerCalendarView>(PART_Calendar);
-
-        DatePickerCalendarView.DateSelectedEvent.AddHandler(OnDateSelected, RoutingStrategies.Bubble, true, _calendar);
-        GotFocusEvent.AddHandler(OnTextBoxGotFocus, _textBox);
-        LostFocusEvent.AddHandler(OnTextBoxLostFocus, _textBox);
-        PointerPressedEvent.AddHandler(OnTextBoxPressed, RoutingStrategies.Tunnel, true, _textBox);
-        SyncSelectedDateToText(SelectedDate);
-    }
-
-    private void OnTextBoxPressed(object? sender, PointerPressedEventArgs e) => InitializePopupOpen();
-
-    private void OnTextBoxGotFocus(object? sender, FocusChangedEventArgs e) => InitializePopupOpen();
-
-    private void OnTextBoxLostFocus(object? sender, FocusChangedEventArgs e) => CommitInput();
-
-    private void InitializePopupOpen()
-    {
-        SetCurrentValue(IsDropdownOpenProperty, true);
-        SetCalendarContextDate();
-        var selectedDate = SelectedDate.HasValue ? ToDateOnly(SelectedDate.Value) : null;
-        _calendar?.MarkDates(selectedDate, selectedDate);
-    }
-
-    private void SetCalendarContextDate()
-    {
-        var date = (SelectedDate.HasValue ? ToDateOnly(SelectedDate.Value) : null) ?? GetToday();
-        _calendar?.SyncContextDate(new DatePickerCalendarContext(date.Year, date.Month));
-        _calendar?.UpdateDayButtons();
-    }
-
-    private void OnDateSelected(object? sender, DatePickerCalendarDayButtonEventArgs e)
-    {
-        SetCurrentValue(SelectedDateProperty, e.Date.HasValue ? (T?)FromDateOnly(e.Date.Value) : null);
-        SetCurrentValue(IsDropdownOpenProperty, false);
-    }
-
-    private void SyncSelectedDateToText(T? value)
-    {
-        if (value is null)
+        if (SelectedDate is null)
         {
             _textBox?.SetValue(TextBox.TextProperty, null);
             _calendar?.ClearSelection();
         }
         else
         {
-            _textBox?.SetValue(TextBox.TextProperty, Format(value.Value, DisplayFormat ?? DEFAULT_DATE_DISPLAY_FORMAT));
-            var dateOnly = ToDateOnly(value.Value);
+            _textBox?.SetValue(TextBox.TextProperty, Format(SelectedDate.Value, DisplayFormat ?? DEFAULT_DATE_DISPLAY_FORMAT));
+            var dateOnly = ToDateOnly(SelectedDate.Value);
             if (dateOnly.HasValue)
                 _calendar?.MarkDates(startDate: dateOnly.Value, endDate: dateOnly.Value);
         }
     }
 
-    /// <inheritdoc/>
-    protected override void OnPointerPressed(PointerPressedEventArgs e)
-    {
-        base.OnPointerPressed(e);
-        if (!e.Handled && e.Source is Visual source)
-        {
-            if (_popup?.IsInsidePopup(source) == true)
-                e.Handled = true;
-            else
-                _textBox?.Focus();
-        }
-    }
-
-    protected override void OnKeyDown(KeyEventArgs e)
-    {
-        switch (e.Key)
-        {
-            case Key.Escape:
-                SetCurrentValue(IsDropdownOpenProperty, false);
-                e.Handled = true;
-                return;
-            case Key.Down:
-                SetCurrentValue(IsDropdownOpenProperty, true);
-                e.Handled = true;
-                return;
-            case Key.Tab:
-                SetCurrentValue(IsDropdownOpenProperty, false);
-                return;
-            case Key.Enter:
-                SetCurrentValue(IsDropdownOpenProperty, false);
-                CommitInput();
-                e.Handled = true;
-                return;
-            default:
-                base.OnKeyDown(e);
-                break;
-        }
-    }
-
-    public void Clear() => SetCurrentValue(SelectedDateProperty, (T?)null);
-
-    protected override void OnLostFocus(FocusChangedEventArgs e)
-    {
-        base.OnLostFocus(e);
-        var element = e.NewFocusedElement;
-        if (Equals(element, _textBox)) return;
-        if (element is Visual v && _popup?.IsInsidePopup(v) == true) return;
-        CommitInput();
-        SetCurrentValue(IsDropdownOpenProperty, false);
-    }
-
-    private void CommitInput()
+    protected override void CommitInput()
     {
         if (string.IsNullOrWhiteSpace(_textBox?.Text))
         {
@@ -196,6 +83,20 @@ public abstract class DatePickerBase<T> : DatePickerBase, IClearControl where T 
             _calendar?.ClearSelection();
         }
     }
+
+    protected override void OnCalendarDateSelected(DateOnly? date)
+    {
+        SetCurrentValue(SelectedDateProperty, date.HasValue ? (T?)FromDateOnly(date.Value) : null);
+    }
+
+    protected override void OnPropertyChanged(AvaloniaPropertyChangedEventArgs change)
+    {
+        base.OnPropertyChanged(change);
+        if (change.Property == SelectedDateProperty)
+            SyncToUI();
+    }
+
+    public void Clear() => SetCurrentValue(SelectedDateProperty, (T?)null);
 
     protected override void UpdateDataValidation(AvaloniaProperty property, BindingValueType state, Exception? error)
     {
